Refactor MainActivity.kt to streamline code and remove unused components
- Removed the CalendarViewModel and associated logic from MainActivity.kt to simplify the structure.
- Eliminated unused imports and commented-out code to enhance readability.
- Retained essential composables for calendar display while ensuring a cleaner codebase.
- Updated comments to reflect the current functionality of the application.
